Start building out the module build logic

+++ BlazarBase/src/main/java/com/hubspot/blazar/base/ModuleBuild.java
@@ -77,7 +77,7 @@ public class ModuleBuild {
     return repoBuildId;
   }
 
-  public long getModuleId() {
+  public int getModuleId() {
     return moduleId;
   }
